refactor: extract duplicated E2E test helpers to conftest.py
Consolidate shared helpers (_with_server, _get_text, _create_proxy,
_chat_body, SSE helpers, fake key builders) into tests/e2e/conftest.py.
Pin anyio_backend to asyncio-only (project doesn't use trio).
Fix unused imports and move type-only imports into TYPE_CHECKING blocks.

Reduces duplication across 15 E2E test files. All helpers are
available as both clean names (get_text) and underscore aliases
(_get_text) for backward compatibility.

1417 tests passing (was 1435; 18 removed trio duplicate test variants).
