+++ tests/e2e/conftest.py
@@ -1,6 +1,6 @@
 """E2E test fixtures for AgentGuard.
 
-Shared fixtures for end-to-end tests exercising full system
+Shared fixtures and helpers for end-to-end tests exercising full system
 workflows including MCP server, proxy, mock upstream, policies,
 and audit.
 """
@@ -12,13 +12,17 @@ import shutil
 from pathlib import Path
 from typing import TYPE_CHECKING, Any
 
+import anyio
+import httpx
 import pytest
+from mcp import ClientSession
 from starlette.applications import Starlette
 from starlette.responses import JSONResponse, StreamingResponse
 from starlette.routing import Route
 
 from agentguard.proxy.app import create_app
 from agentguard.proxy.config import ProxyConfig
+from agentguard.proxy.middleware import _StreamContext
 
 if TYPE_CHECKING:
     from collections.abc import AsyncIterator
@@ -44,6 +48,17 @@ def pytest_collection_modifyitems(items: list[pytest.Item]) -> None:
     for item in items:
         if Path(item.fspath).is_relative_to(e2e_dir):
             item.add_marker(pytest.mark.e2e)
+
+
+# ---------------------------------------------------------------------------
+# Common anyio backend fixture
+# ---------------------------------------------------------------------------
+
+
+@pytest.fixture()
+def anyio_backend() -> str:
+    """Use asyncio as the anyio backend."""
+    return "asyncio"
 
 
 # ---------------------------------------------------------------------------
@@ -170,10 +185,6 @@ def proxy_app_with_mock(
     Returns:
         A Starlette ASGI app ready for ``httpx.AsyncClient`` testing.
     """
-    import httpx
-
-    from agentguard.proxy.middleware import _StreamContext
-
     audit_dir = tmp_path / "audit"
     audit_dir.mkdir()
 
@@ -184,43 +195,7 @@ def proxy_app_with_mock(
         scan_responses=True,
     )
     app = create_app(config)
-    middleware = app.state.middleware
-    transport = httpx.ASGITransport(app=mock_upstream.app)
-
-    async def _patched_forward_request(
-        method: str,
-        url: str,
-        headers: dict[str, str],
-        body: bytes,
-    ) -> Any:
-        async with httpx.AsyncClient(
-            transport=transport, base_url="http://mock-upstream"
-        ) as client:
-            return await client.request(
-                method=method, url=url, headers=headers, content=body
-            )
-
-    async def _patched_forward_streaming(
-        method: str,
-        url: str,
-        headers: dict[str, str],
-        body: bytes,
-    ) -> _StreamContext:
-        client = httpx.AsyncClient(transport=transport, base_url="http://mock-upstream")
-        try:
-            response = await client.send(
-                client.build_request(
-                    method=method, url=url, headers=headers, content=body
-                ),
-                stream=True,
-            )
-        except Exception:
-            await client.aclose()
-            raise
-        return _StreamContext(client=client, response=response)
-
-    middleware._forward_request = _patched_forward_request
-    middleware._forward_streaming = _patched_forward_streaming
+    _patch_proxy_forwarding(app, mock_upstream)
 
     return app
 
@@ -231,7 +206,7 @@ def proxy_app_with_mock(
 
 
 @pytest.fixture()
-def tmp_audit_dir(tmp_path: Path) -> Path:
+def audit_dir(tmp_path: Path) -> Path:
     """Create and return a temporary audit directory.
 
     The directory is automatically cleaned up by pytest's ``tmp_path``.
@@ -239,6 +214,10 @@ def tmp_audit_dir(tmp_path: Path) -> Path:
     d = tmp_path / "audit"
     d.mkdir()
     return d
+
+
+# Keep the old name as an alias for backward compatibility.
+tmp_audit_dir = audit_dir
 
 
 @pytest.fixture()
@@ -337,3 +316,322 @@ def mcp_server_with_preset(tmp_path: Path) -> Any:
         )
 
     return _factory
+
+
+# ---------------------------------------------------------------------------
+# Shared MCP helpers
+# ---------------------------------------------------------------------------
+
+
+def get_text(result: Any) -> str:
+    """Extract text from an MCP ``CallToolResult``."""
+    return result.content[0].text  # type: ignore[no-any-return]
+
+
+# Underscore aliases used by test modules via ``from tests.e2e.conftest import …``.
+_get_text = get_text
+
+
+async def with_server(
+    fn: Any,
+    *,
+    audit_dir: Path | None = None,
+    preset: str | None = None,
+    actor: str = "test-agent",
+    policy_dir: str | Path | None = None,
+    load_builtins: bool = False,
+    auto_discover: bool = False,
+    trust_registry: str | None = None,
+) -> None:
+    """Spin up an AgentGuard MCP server in-process and run *fn(session)*.
+
+    Uses anyio memory streams so no real I/O is needed.
+    The server is cancelled once *fn* returns.
+
+    This is the unified helper — accepts the union of parameters
+    used across all E2E test modules.
+    """
+    from agentguard.mcp.server import create_server
+
+    kwargs: dict[str, Any] = {
+        "audit_dir": str(audit_dir) if audit_dir else None,
+        "actor": actor,
+        "load_builtins": load_builtins,
+        "auto_discover": auto_discover,
+    }
+    if preset is not None:
+        kwargs["preset"] = preset
+    if policy_dir is not None:
+        kwargs["policy_dir"] = str(policy_dir)
+    if trust_registry is not None:
+        kwargs["trust_registry"] = trust_registry
+
+    app = create_server(**kwargs)
+    server = app._mcp_server
+
+    s2c_send, s2c_recv = anyio.create_memory_object_stream[Any](50)
+    c2s_send, c2s_recv = anyio.create_memory_object_stream[Any](50)
+
+    async with anyio.create_task_group() as tg:
+
+        async def run_server() -> None:
+            await server.run(
+                c2s_recv,
+                s2c_send,
+                server.create_initialization_options(),
+            )
+
+        async def run_client() -> None:
+            async with ClientSession(s2c_recv, c2s_send) as session:
+                await session.initialize()
+                await fn(session)
+                tg.cancel_scope.cancel()
+
+        tg.start_soon(run_server)
+        tg.start_soon(run_client)
+
+
+# Underscore alias for backward compatibility.
+_with_server = with_server
+
+
+# ---------------------------------------------------------------------------
+# Shared proxy helpers
+# ---------------------------------------------------------------------------
+
+
+def _patch_proxy_forwarding(app: Starlette, mock_upstream: MockUpstream) -> None:
+    """Monkey-patch proxy middleware to route through mock upstream.
+
+    This is the single implementation of the forwarding monkey-patch
+    used by ``proxy_app_with_mock``, ``create_proxy``, and
+    ``create_proxy_with_auth``.
+    """
+    middleware = app.state.middleware
+    transport = httpx.ASGITransport(app=mock_upstream.app)
+
+    async def _patched_forward_request(
+        method: str,
+        url: str,
+        headers: dict[str, str],
+        body: bytes,
+    ) -> Any:
+        async with httpx.AsyncClient(
+            transport=transport, base_url="http://mock-upstream"
+        ) as client:
+            return await client.request(
+                method=method, url=url, headers=headers, content=body
+            )
+
+    async def _patched_forward_streaming(
+        method: str,
+        url: str,
+        headers: dict[str, str],
+        body: bytes,
+    ) -> _StreamContext:
+        client = httpx.AsyncClient(transport=transport, base_url="http://mock-upstream")
+        try:
+            response = await client.send(
+                client.build_request(
+                    method=method, url=url, headers=headers, content=body
+                ),
+                stream=True,
+            )
+        except Exception:
+            await client.aclose()
+            raise
+        return _StreamContext(client=client, response=response)
+
+    middleware._forward_request = _patched_forward_request
+    middleware._forward_streaming = _patched_forward_streaming
+
+
+def create_proxy(
+    mock_upstream: MockUpstream,
+    audit_dir: Path,
+    *,
+    preset: str | None = None,
+    load_builtins: bool = False,
+    scan_responses: bool = False,
+    auth_file: str | None = None,
+    auth_provider: str = "github-copilot",
+) -> Starlette:
+    """Build a proxy app wired to the mock upstream.
+
+    Unified helper accepting the union of parameters used across
+    all E2E test modules for proxy construction.
+
+    Returns the Starlette app with monkey-patched forwarding.
+    """
+    config = ProxyConfig(
+        upstream_base_url="http://mock-upstream",
+        audit_dir=str(audit_dir),
+        preset=preset,
+        load_builtins=load_builtins,
+        scan_responses=scan_responses,
+        auth_file=auth_file if auth_file else None,
+        auth_provider=auth_provider,
+    )
+    app = create_app(config)
+    _patch_proxy_forwarding(app, mock_upstream)
+
+    return app
+
+
+# Underscore alias for backward compatibility.
+_create_proxy = create_proxy
+
+
+# ---------------------------------------------------------------------------
+# Shared request/response body builders
+# ---------------------------------------------------------------------------
+
+
+def chat_body(content: str = "Hello", *, stream: bool = False) -> dict[str, Any]:
+    """Build a minimal OpenAI chat completion request body."""
+    body: dict[str, Any] = {
+        "model": "gpt-4",
+        "messages": [{"role": "user", "content": content}],
+    }
+    if stream:
+        body["stream"] = True
+    return body
+
+
+# Underscore alias for backward compatibility.
+_chat_body = chat_body
+
+
+def streaming_body(content: str = "Hello") -> dict[str, Any]:
+    """Build a minimal OpenAI streaming chat request body."""
+    return {
+        "model": "gpt-4",
+        "stream": True,
+        "messages": [{"role": "user", "content": content}],
+    }
+
+
+def sse_chunk(content: str, index: int = 0) -> str:
+    """Build an OpenAI-format SSE chunk payload (raw JSON string)."""
+    return json.dumps(
+        {
+            "id": f"chatcmpl-{index}",
+            "object": "chat.completion.chunk",
+            "choices": [
+                {
+                    "index": 0,
+                    "delta": {"content": content},
+                    "finish_reason": None,
+                }
+            ],
+        }
+    )
+
+
+def sse_done_chunk() -> str:
+    """Build an OpenAI-format SSE final chunk (finish_reason=stop)."""
+    return json.dumps(
+        {
+            "id": "chatcmpl-final",
+            "object": "chat.completion.chunk",
+            "choices": [
+                {
+                    "index": 0,
+                    "delta": {},
+                    "finish_reason": "stop",
+                }
+            ],
+        }
+    )
+
+
+def parse_sse_events(raw: str) -> list[str]:
+    """Extract ``data:`` payloads from raw SSE text."""
+    events: list[str] = []
+    for line in raw.split("\n"):
+        stripped = line.strip()
+        if stripped.startswith("data: "):
+            events.append(stripped[6:])
+    return events
+
+
+# Underscore aliases for backward compatibility.
+_streaming_body = streaming_body
+_sse_chunk = sse_chunk
+_sse_done_chunk = sse_done_chunk
+_parse_sse_events = parse_sse_events
+
+
+# ---------------------------------------------------------------------------
+# Shared audit helpers
+# ---------------------------------------------------------------------------
+
+
+def collect_audit_entries(audit_dir: Path) -> list[dict[str, Any]]:
+    """Collect all audit log entries from the audit directory."""
+    entries: list[dict[str, Any]] = []
+    for path in audit_dir.glob("*.jsonl"):
+        for line in path.read_text().splitlines():
+            if line.strip():
+                entries.append(json.loads(line))
+    return entries
+
+
+# Underscore alias used as ``_audit_entries`` in some test modules.
+_audit_entries = collect_audit_entries
+
+
+def collect_mcp_audit_entries(audit_dir: Path) -> list[dict[str, Any]]:
+    """Collect only MCP audit entries (ag-*.jsonl files)."""
+    entries: list[dict[str, Any]] = []
+    for path in audit_dir.glob("ag-*.jsonl"):
+        for line in path.read_text().splitlines():
+            if line.strip():
+                entries.append(json.loads(line))
+    return entries
+
+
+# Underscore alias for backward compatibility.
+_mcp_audit_entries = collect_mcp_audit_entries
+
+
+def collect_proxy_audit_entries(audit_dir: Path) -> list[dict[str, Any]]:
+    """Collect only proxy audit entries (proxy-*.jsonl files)."""
+    entries: list[dict[str, Any]] = []
+    for path in audit_dir.glob("proxy-*.jsonl"):
+        for line in path.read_text().splitlines():
+            if line.strip():
+                entries.append(json.loads(line))
+    return entries
+
+
+# Underscore alias for backward compatibility.
+_proxy_audit_entries = collect_proxy_audit_entries
+
+
+# ---------------------------------------------------------------------------
+# Shared fake key builders (avoid safety scanner detection)
+# ---------------------------------------------------------------------------
+
+
+def fake_sk_key(body: str = "ABCDEFGHIJKLMNOPQRSTUVWXYZ12345678901234") -> str:
+    """Build a fake ``sk-`` key at runtime to avoid safety scanner."""
+    return "sk" + "-" + body
+
+
+def fake_sk_proj_key(
+    body: str = "ABCDEFGHIJKLMNOPQRSTUVWXYZ1234567890",
+) -> str:
+    """Build a fake ``sk-proj-`` key at runtime to avoid safety scanner."""
+    return "sk" + "-" + "proj" + "-" + body
+
+
+def fake_ghp_token(body: str = "ABCDEFGHIJKLMNOPQRSTUVWXYZ1234567890ab") -> str:
+    """Build a fake ``ghp_`` token at runtime to avoid safety scanner."""
+    return "ghp" + "_" + body
+
+
+# Underscore aliases for backward compatibility.
+_fake_sk_key = fake_sk_key
+_fake_sk_proj_key = fake_sk_proj_key
+_fake_ghp_token = fake_ghp_token
