+++ tests/e2e/test_error_handling.py
@@ -21,165 +21,23 @@ from __future__ import annotations
 import json
 import os
 import sys
-from typing import TYPE_CHECKING, Any
+from typing import TYPE_CHECKING
 
-import anyio
 import httpx
 import pytest
-from mcp import ClientSession
 
-from agentguard.proxy.app import create_app
-from agentguard.proxy.config import ProxyConfig
-from agentguard.proxy.middleware import _StreamContext
+from tests.e2e.conftest import (
+    _create_proxy,
+    _get_text,
+    _with_server,
+)
 
 if TYPE_CHECKING:
     from pathlib import Path
 
+    from mcp import ClientSession
+
     from tests.e2e.conftest import MockUpstream
-
-
-# ---------------------------------------------------------------------------
-# Fixtures
-# ---------------------------------------------------------------------------
-
-
-@pytest.fixture()
-def anyio_backend() -> str:
-    """Use asyncio as the anyio backend."""
-    return "asyncio"
-
-
-@pytest.fixture()
-def audit_dir(tmp_path: Path) -> Path:
-    """Create a temp directory for audit logs."""
-    d = tmp_path / "audit"
-    d.mkdir()
-    return d
-
-
-# ---------------------------------------------------------------------------
-# Helpers
-# ---------------------------------------------------------------------------
-
-
-def _get_text(result: Any) -> str:
-    """Extract text from an MCP CallToolResult."""
-    return result.content[0].text  # type: ignore[no-any-return]
-
-
-async def _with_server(
-    fn: Any,
-    *,
-    audit_dir: Path | None = None,
-    actor: str = "test-agent",
-    builtins: bool = False,
-    preset: str | None = None,
-    policy_dir: Path | None = None,
-) -> None:
-    """Spin up an AgentGuard MCP server in-process and run *fn(session)*.
-
-    Uses anyio memory streams — no real I/O needed.
-    """
-    from agentguard.mcp.server import create_server
-
-    kwargs: dict[str, Any] = {
-        "audit_dir": str(audit_dir) if audit_dir else None,
-        "actor": actor,
-        "load_builtins": builtins,
-    }
-    if preset is not None:
-        kwargs["preset"] = preset
-    if policy_dir is not None:
-        kwargs["policy_dir"] = str(policy_dir)
-
-    app = create_server(**kwargs)
-    server = app._mcp_server
-
-    s2c_send, s2c_recv = anyio.create_memory_object_stream[Any](50)
-    c2s_send, c2s_recv = anyio.create_memory_object_stream[Any](50)
-
-    async with anyio.create_task_group() as tg:
-
-        async def run_server() -> None:
-            await server.run(
-                c2s_recv,
-                s2c_send,
-                server.create_initialization_options(),
-            )
-
-        async def run_client() -> None:
-            async with ClientSession(s2c_recv, c2s_send) as session:
-                await session.initialize()
-                await fn(session)
-                tg.cancel_scope.cancel()
-
-        tg.start_soon(run_server)
-        tg.start_soon(run_client)
-
-
-def _create_proxy(
-    mock_upstream: MockUpstream,
-    tmp_path: Path,
-    *,
-    preset: str | None = None,
-    load_builtins: bool = False,
-    scan_responses: bool = False,
-) -> Any:
-    """Build a proxy app wired to the mock upstream.
-
-    Returns the Starlette app with forwarding methods monkey-patched
-    to route through ASGI transport pointing at the mock.
-    """
-    ad = tmp_path / "audit"
-    ad.mkdir(exist_ok=True)
-
-    config = ProxyConfig(
-        upstream_base_url="http://mock-upstream",
-        audit_dir=str(ad),
-        preset=preset,
-        load_builtins=load_builtins,
-        scan_responses=scan_responses,
-    )
-    app = create_app(config)
-    middleware = app.state.middleware
-    transport = httpx.ASGITransport(app=mock_upstream.app)
-
-    async def _patched_forward_request(
-        method: str,
-        url: str,
-        headers: dict[str, str],
-        body: bytes,
-    ) -> Any:
-        async with httpx.AsyncClient(
-            transport=transport, base_url="http://mock-upstream"
-        ) as client:
-            return await client.request(
-                method=method, url=url, headers=headers, content=body
-            )
-
-    async def _patched_forward_streaming(
-        method: str,
-        url: str,
-        headers: dict[str, str],
-        body: bytes,
-    ) -> _StreamContext:
-        client = httpx.AsyncClient(transport=transport, base_url="http://mock-upstream")
-        try:
-            response = await client.send(
-                client.build_request(
-                    method=method, url=url, headers=headers, content=body
-                ),
-                stream=True,
-            )
-        except Exception:
-            await client.aclose()
-            raise
-        return _StreamContext(client=client, response=response)
-
-    middleware._forward_request = _patched_forward_request
-    middleware._forward_streaming = _patched_forward_streaming
-
-    return app
 
 
 # ===========================================================================
@@ -474,10 +332,10 @@ class TestProxyMalformedInput:
 
     @pytest.mark.anyio()
     async def test_sg_13_6_malformed_json_no_crash(
-        self, mock_upstream: MockUpstream, tmp_path: Path
+        self, mock_upstream: MockUpstream, audit_dir: Path
     ) -> None:
         """Proxy does not crash on malformed JSON request body."""
-        app = _create_proxy(mock_upstream, tmp_path)
+        app = _create_proxy(mock_upstream, audit_dir)
 
         async with httpx.AsyncClient(
             transport=httpx.ASGITransport(app=app),
@@ -494,11 +352,11 @@ class TestProxyMalformedInput:
 
     @pytest.mark.anyio()
     async def test_sg_13_6_malformed_json_passes_through(
-        self, mock_upstream: MockUpstream, tmp_path: Path
+        self, mock_upstream: MockUpstream, audit_dir: Path
     ) -> None:
         """Malformed JSON is forwarded to upstream without scanning."""
         mock_upstream.set_response({"choices": [{"message": {"content": "ok"}}]})
-        app = _create_proxy(mock_upstream, tmp_path, load_builtins=True)
+        app = _create_proxy(mock_upstream, audit_dir, load_builtins=True)
 
         async with httpx.AsyncClient(
             transport=httpx.ASGITransport(app=app),
@@ -514,11 +372,11 @@ class TestProxyMalformedInput:
 
     @pytest.mark.anyio()
     async def test_sg_13_7_empty_body_no_crash(
-        self, mock_upstream: MockUpstream, tmp_path: Path
+        self, mock_upstream: MockUpstream, audit_dir: Path
     ) -> None:
         """Proxy does not crash on empty request body."""
         mock_upstream.set_response({"choices": [{"message": {"content": "ok"}}]})
-        app = _create_proxy(mock_upstream, tmp_path)
+        app = _create_proxy(mock_upstream, audit_dir)
 
         async with httpx.AsyncClient(
             transport=httpx.ASGITransport(app=app),
@@ -533,11 +391,11 @@ class TestProxyMalformedInput:
 
     @pytest.mark.anyio()
     async def test_sg_13_7_empty_body_audited(
-        self, mock_upstream: MockUpstream, tmp_path: Path
+        self, mock_upstream: MockUpstream, audit_dir: Path
     ) -> None:
         """Empty-body request is recorded in the proxy audit log."""
         mock_upstream.set_response({"choices": [{"message": {"content": "ok"}}]})
-        app = _create_proxy(mock_upstream, tmp_path)
+        app = _create_proxy(mock_upstream, audit_dir)
 
         async with httpx.AsyncClient(
             transport=httpx.ASGITransport(app=app),
@@ -550,8 +408,7 @@ class TestProxyMalformedInput:
             )
 
         # Check audit log exists
-        ad = tmp_path / "audit"
-        jsonl_files = list(ad.glob("*.jsonl"))
+        jsonl_files = list(audit_dir.glob("*.jsonl"))
         assert len(jsonl_files) >= 1, "Expected audit log for empty-body request"
