+++ tests/e2e/test_policy_engine.py
@@ -15,19 +15,20 @@ Tests cover:
 from __future__ import annotations
 
 import json
-from typing import TYPE_CHECKING, Any
+from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
     from pathlib import Path
 
-import anyio
+    from mcp import ClientSession
+
 import pytest
-from mcp import ClientSession
 
 from agentguard.policies.discovery import discover_policies
 from agentguard.policies.guard import Guard
 from agentguard.policies.loader import load_policy_from_string
 from agentguard.policies.presets import PRESET_POLICIES, Preset
+from tests.e2e.conftest import _get_text, _with_server
 
 # ---------------------------------------------------------------------------
 # Helpers
@@ -73,75 +74,6 @@ def _write_policy_file(directory: Path, filename: str, content: str) -> Path:
     path = directory / filename
     path.write_text(content)
     return path
-
-
-def _get_text(result: Any) -> str:
-    """Extract text from an MCP CallToolResult."""
-    return result.content[0].text  # type: ignore[no-any-return]
-
-
-async def _with_server(
-    fn: Any,
-    *,
-    audit_dir: Path | None = None,
-    preset: str | None = None,
-    actor: str = "test-agent",
-    policy_dir: str | None = None,
-    auto_discover: bool = False,
-    load_builtins: bool = False,
-) -> None:
-    """Spin up an AgentGuard MCP server and run *fn(session)*.
-
-    Uses anyio memory streams so no real I/O is needed.
-    The server is cancelled once *fn* returns.
-    """
-    from agentguard.mcp.server import create_server
-
-    app = create_server(
-        preset=preset,
-        audit_dir=str(audit_dir) if audit_dir else None,
-        actor=actor,
-        policy_dir=policy_dir,
-        auto_discover=auto_discover,
-        load_builtins=load_builtins,
-    )
-
-    server = app._mcp_server
-
-    s2c_send, s2c_recv = anyio.create_memory_object_stream[Any](50)
-    c2s_send, c2s_recv = anyio.create_memory_object_stream[Any](50)
-
-    async with anyio.create_task_group() as tg:
-
-        async def run_server() -> None:
-            await server.run(
-                c2s_recv,
-                s2c_send,
-                server.create_initialization_options(),
-            )
-
-        async def run_client() -> None:
-            async with ClientSession(s2c_recv, c2s_send) as session:
-                await session.initialize()
-                await fn(session)
-                tg.cancel_scope.cancel()
-
-        tg.start_soon(run_server)
-        tg.start_soon(run_client)
-
-
-@pytest.fixture()
-def anyio_backend() -> str:
-    """Use asyncio as the anyio backend."""
-    return "asyncio"
-
-
-@pytest.fixture()
-def audit_dir(tmp_path: Path) -> Path:
-    """Create a temp directory for audit logs."""
-    d = tmp_path / "audit"
-    d.mkdir()
-    return d
 
 
 # ---------------------------------------------------------------------------
